iteration4batch
Sadly, due to the lack of time between iteration 3 and this one, not an enormous amount of tangible progress was made -- however, instead of getting a jumpstart on charting goals for DSAM 2 and other social media outlets for it, I moved forward with some small goals.

1. Expand upon roles within church and parameters of the church itself.

2. Begin sourcing historical information on previous Popes -- especially those alive to actively push back against internet connectivity.

3. Map out the most popular/influential videos and visualize trends.

4. Continue to write in my journal, review it, and prepare for the finale.

A piece of commentary: Even though I got nowhere near I thought I'd be in this semester, even with all of my barriers -- it was interesting seeing how less panicked my journal entries and annotated reading notes got throughout the semester.